Updated comments and made small corrections
Fixed some typos in the comments. Added a comment in the DS1 beat classification file to note that the same file can be used to generate results for DS2 with some editing; also added a line to save results. Fixed the flag for plotATM.m incorporation in MIT_CUR_matrix_constructor.m and fixed an initialization that was previously unused.